refactor: ui 패키지 정리
- widget을 ui로 이동
- main안으로 friendList 패키지 이동
